requirements.txt
ably==2.0.12
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
boto3-stubs==1.38.21
botocore-stubs==1.38.19
cachetools==5.5.2
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
cohere==5.14.2
coloredlogs==15.0.1
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
distlib==0.3.9
distro==1.9.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
fastavro==1.10.0
fastembed==0.6.1
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
hyperframe==6.1.0
identify==2.6.9
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-google-genai==2.1.2
langchain-huggingface==0.1.2
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.8
langsmith==0.3.30
Levenshtein==0.27.1
litellm==1.70.2
loguru==0.7.3
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
methoddispatch==3.0.2
mmh3==5.1.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
mypy-boto3-bedrock-runtime==1.38.4
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx
nodeenv==1.9.1
numpy
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnxruntime
openai==1.75.0
opentelemetry-api==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opik==1.7.25
orjson==3.10.16
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
portalocker==2.10.1
pre_commit==4.2.0
prompt_toolkit==3.0.50
promptlayer==1.0.48
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py_rust_stemmers==0.1.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
pyee==12.1.1
Pygments==2.19.1
PyMySQL==1.1.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
pyzmq==26.4.0
qdrant-client==1.13.3
RapidFuzz==3.13.0
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy
sentence-transformers==4.0.2
sentry-sdk==2.29.1
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.46.1
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.2
triton==3.2.0
types-awscrt==0.27.2
types-requests==2.32.0.20250328
types-s3transfer==0.12.0
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.4.0
uuid6==2024.7.10
uvicorn==0.34.0
virtualenv==20.30.0
wcwidth==0.2.13
websockets==13.1
wrapt==1.17.2
yarl==1.19.0
zipp==3.21.0
zstandard==0.23.0
